2.1.0 - beta 3 SelectCalendarsActivity: -fixed calendar list not scrolling when all calendars are set to sync -now uses same imageview sidebar as the event list
DatabaseInterface:
-fixed regression that caused event colors not to be read properly
-fixed bug where event calendar ids were being stored as ints instead of longs

Prefs:
-fixed bug where restoreDefaultPreferences() did not restore all values properly

EventSilencerService:
-fixed bug in shouldEventSilence() where events were improperly flagged not to silence

+++ tacere/src/main/java/org/ciasaboark/tacere/database/SimpleCalendarEvent.java
@@ -15,7 +15,7 @@ public class SimpleCalendarEvent {
     public static final long MILLISECONDS_IN_DAY = MILLISECONDS_IN_MINUTE * 60 * 24;
     @SuppressWarnings("unused")
     private static final String TAG = "CalEvent";
-    private int calendarId;
+    private long calendarId;
     private int instanceId;     //instance ids are unique in repeating events
     private int eventId;        //event ids are shared among repeating events
     private String title;
@@ -28,7 +28,7 @@ public class SimpleCalendarEvent {
     private boolean isAllDay;
 
 
-    public SimpleCalendarEvent(int calendarId, int instanceId, int eventId, String title, long begin, long end,
+    public SimpleCalendarEvent(long calendarId, int instanceId, int eventId, String title, long begin, long end,
                                String description, int displayColor, boolean isFreeTime, boolean isAllDay) {
         this.calendarId = calendarId;
         this.instanceId = instanceId;
@@ -55,7 +55,7 @@ public class SimpleCalendarEvent {
         return dateFormatter.format(date);
     }
 
-    public int getCalendarId() {
+    public long getCalendarId() {
         return calendarId;
     }
